refactor: 도메인에서 발생 가능한 예외 클래스들의 패키지 이동
