+++ src/main/java/chess/domain/exception/DomainException.java
@@ -1,4 +1,6 @@
-package chess.exception;
+package chess.domain.exception;
+
+import chess.exception.HandledException;
 
 public class DomainException extends HandledException {
     public DomainException(String message) {
