+++ src/main/java/chess/domain/exception/InvalidMoveStrategyException.java
@@ -1,4 +1,6 @@
-package chess.exception;
+package chess.domain.exception;
+
+import chess.exception.HandledException;
 
 public class InvalidMoveStrategyException extends HandledException {
     public InvalidMoveStrategyException(String message) {
