+++ src/main/java/chess/domain/exception/InvalidStateException.java
@@ -1,4 +1,6 @@
-package chess.exception;
+package chess.domain.exception;
+
+import chess.exception.HandledException;
 
 public class InvalidStateException extends HandledException {
     public InvalidStateException(String message) {
