+++ src/main/java/chess/domain/exception/InvalidTurnException.java
@@ -1,4 +1,6 @@
-package chess.exception;
+package chess.domain.exception;
+
+import chess.exception.HandledException;
 
 public class InvalidTurnException extends HandledException {
     public InvalidTurnException(String message) {
