+++ src/main/java/chess/domain/exception/NoSuchDirectionException.java
@@ -1,4 +1,6 @@
-package chess.exception;
+package chess.domain.exception;
+
+import chess.exception.HandledException;
 
 public class NoSuchDirectionException extends HandledException {
     public NoSuchDirectionException(String message) {
