+++ src/main/java/chess/domain/exception/NullObjectSelectionException.java
@@ -1,4 +1,6 @@
-package chess.exception;
+package chess.domain.exception;
+
+import chess.exception.HandledException;
 
 public class NullObjectSelectionException extends HandledException {
     public NullObjectSelectionException(String message) {
