+++ src/main/java/chess/domain/exception/SamePositionChosenException.java
@@ -1,4 +1,6 @@
-package chess.exception;
+package chess.domain.exception;
+
+import chess.exception.HandledException;
 
 public class SamePositionChosenException extends HandledException {
     public SamePositionChosenException(String message) {
