+++ src/main/java/chess/domain/piece/Color.java
@@ -1,6 +1,6 @@
 package chess.domain.piece;
 
-import chess.exception.DomainException;
+import chess.domain.exception.DomainException;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
