+++ src/main/java/chess/domain/piece/strategy/KingMoveStrategy.java
@@ -2,7 +2,7 @@ package chess.domain.piece.strategy;
 
 import chess.domain.order.MoveRoute;
 import chess.domain.position.Direction;
-import chess.exception.InvalidMoveStrategyException;
+import chess.domain.exception.InvalidMoveStrategyException;
 
 public class KingMoveStrategy extends DefaultMoveStrategy {
     public static final int KING_MOVABLE_ROUTE_LENGTH = 2;
