+++ src/main/java/chess/domain/piece/strategy/PawnMoveStrategy.java
@@ -4,7 +4,7 @@ import chess.domain.order.MoveRoute;
 import chess.domain.piece.Color;
 import chess.domain.position.Direction;
 import chess.domain.position.Rank;
-import chess.exception.InvalidMoveStrategyException;
+import chess.domain.exception.InvalidMoveStrategyException;
 
 public class PawnMoveStrategy extends DefaultMoveStrategy {
     public static final int PAWN_MAXIMUM_MOVABLE_ROUTE_LENGTH = 3;
