+++ src/main/java/chess/domain/position/Rank.java
@@ -1,6 +1,6 @@
 package chess.domain.position;
 
-import chess.exception.DomainException;
+import chess.domain.exception.DomainException;
 
 import java.util.*;
 import java.util.function.Function;
