+++ src/main/java/chess/domain/state/GameEnd.java
@@ -1,6 +1,6 @@
 package chess.domain.state;
 
-import chess.exception.InvalidStateException;
+import chess.domain.exception.InvalidStateException;
 
 public class GameEnd implements State {
     public GameEnd() {
