+++ src/test/java/chess/domain/position/PositionTest.java
@@ -1,6 +1,6 @@
 package chess.domain.position;
 
-import chess.exception.DomainException;
+import chess.domain.exception.DomainException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
